Fix ScrollView zoom scale assertion warning.
Before the view is laid out, calculations of zoom scale result in all 0 because the view's frame is currently `.zero`, this then sets ScrollView's zoom to 0 and triggers internal assertion and prints a warning. This mainly happens with local images which Nuke finishes loading before view layout. After the view finishes loading and appears on screen, another calculation is done to ensure correct zoom, which is why the final display looks OK. So the first pass can be avoided completely. (It's always wrong anyway.)